requirements.txt
langchain
streamlit
fastapi
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
lancedb